+++ hreq-client/test/Hreq/HttpBin/SuccessSpec.hs
@@ -17,6 +17,10 @@ successSpec = do
        r  <- runHttpBin $ hreq @("get" :> RawResponse GET) Empty
        resStatusCode r `shouldBe` 200
 
+    it "works with specifying status codes where by received code must match expected" $ do
+       r <- runHttpBin $ hreq @(Verb GET '[ ResStatus 200 ]) Empty
+       resStatusCode r `shouldBe` 200
+
     it "works with post requests" $ do
        r  <- runHttpBin
                $ hreq @("post" :> JsonBody String :> PostJson Value) ("foo" :. Empty)