+++ hreq-client/CHANGELOG.md
@@ -1,3 +1,10 @@
-## Hreq-Client 0.1.0.0
+## Hreq-Client 0.1.1.0
+
+* Matching status codes with in response.
+
+One can specify that a certain status code is expected in the http response
+and when its not received we throw an error.
+
+## Hreq-Client 0.1.1.0
 
 * Initial public release